requirements.txt
flask==3.1.0
flask-cors==6.0.0
Pillow==11.0.0
numpy==2.1.3
opencv-python-headless==4.10.0.84
torch==2.5.1
torchvision==0.20.1
grad-cam==1.5.5
ultralytics==8.4.19
gunicorn==23.0.0
transformers>=4.35.0